fix entity and device registry needed to add unique_id
